高频SQL分析 加user
