+++ src/main/java/org/opencloudb/response/ShowSQLHigh.java
@@ -12,7 +12,9 @@ import org.opencloudb.net.mysql.FieldPacket;
 import org.opencloudb.net.mysql.ResultSetHeaderPacket;
 import org.opencloudb.net.mysql.RowDataPacket;
 import org.opencloudb.stat.HighFrequencySqlAnalyzer;
-import org.opencloudb.stat.HighFrequencySqlAnalyzer.SqlFrequency;
+import org.opencloudb.stat.SqlFrequency;
+import org.opencloudb.stat.UserStat;
+import org.opencloudb.stat.UserStatAnalyzer;
 import org.opencloudb.util.LongUtil;
 import org.opencloudb.util.StringUtil;
 
@@ -24,7 +26,7 @@ import org.opencloudb.util.StringUtil;
  */
 public final class ShowSQLHigh {
 	
-	private static final int FIELD_COUNT = 8;
+	private static final int FIELD_COUNT = 9;
     private static final ResultSetHeaderPacket header = PacketUtil.getHeader(FIELD_COUNT);
     private static final FieldPacket[] fields = new FieldPacket[FIELD_COUNT];
     private static final EOFPacket eof = new EOFPacket();
@@ -36,7 +38,10 @@ public final class ShowSQLHigh {
         
         fields[i] = PacketUtil.getField("ID", Fields.FIELD_TYPE_LONGLONG);
         fields[i++].packetId = ++packetId;
-
+        
+        fields[i] = PacketUtil.getField("USER", Fields.FIELD_TYPE_VARCHAR);
+        fields[i++].packetId = ++packetId;
+        
         fields[i] = PacketUtil.getField("FREQUENCY", Fields.FIELD_TYPE_LONGLONG);
         fields[i++].packetId = ++packetId;
         
@@ -73,7 +78,8 @@ public final class ShowSQLHigh {
         buffer = eof.write(buffer, c,true);
 
         // write rows
-        byte packetId = eof.packetId;        
+        byte packetId = eof.packetId;     
+        /*
         List<Map.Entry<String, SqlFrequency>> list = HighFrequencySqlAnalyzer.getInstance().getSqlFrequency( isClear );
         if ( list != null ) {        
 	        for (int i = 0; i < list.size(); i++) {
@@ -83,7 +89,20 @@ public final class ShowSQLHigh {
 	            buffer = row.write(buffer, c,true);
 	        }
         }
-
+        */
+        Map<String, UserStat> statMap = UserStatAnalyzer.getInstance().getUserStatMap();
+    	for (UserStat userStat : statMap.values()) {
+        	String user = userStat.getUser();        
+        	 List<Map.Entry<String, SqlFrequency>> list=userStat.getSqlHigh().getSqlFrequency( isClear );
+             if ( list != null ) {        
+     	        for (int i = 0; i < list.size(); i++) {
+     	        	SqlFrequency sqlFrequency = list.get(i).getValue();
+     	        	RowDataPacket row = getRow(i, user,sqlFrequency.getSql(), sqlFrequency.getCount(), sqlFrequency.getAvgTime(),sqlFrequency.getMaxTime(),sqlFrequency.getMinTime(),sqlFrequency.getExecuteTime(),sqlFrequency.getLastTime(), c.getCharset());
+     	            row.packetId = ++packetId;
+     	            buffer = row.write(buffer, c,true);
+     	        }
+             }
+    	}    
         // write last eof
         EOFPacket lastEof = new EOFPacket();
         lastEof.packetId = ++packetId;
@@ -93,9 +112,10 @@ public final class ShowSQLHigh {
         c.write(buffer);
     }
 
-    private static RowDataPacket getRow(int i, String sql, int count, long avgTime,long maxTime,long minTime,long executTime,long lastTime, String charset) {
+    private static RowDataPacket getRow(int i, String user,String sql, int count, long avgTime,long maxTime,long minTime,long executTime,long lastTime, String charset) {
         RowDataPacket row = new RowDataPacket(FIELD_COUNT);
         row.add( LongUtil.toBytes( i ) );
+        row.add( StringUtil.encode( user, charset) );
         row.add( LongUtil.toBytes( count ) );
         row.add( LongUtil.toBytes( avgTime ) );
         row.add( LongUtil.toBytes( maxTime ) );
