+++ src/main/java/org/opencloudb/stat/UserStat.java
@@ -30,6 +30,7 @@ public class UserStat {
 	private String user;
 	
 
+	private UserSqlHigh sqlHighStat = null;
 	
 	public UserStat(String user) {
 		super();
@@ -37,6 +38,7 @@ public class UserStat {
 		this.rwStat = new UserRWStat();
 		this.sqlStat = new UserSqlStat(50);
 		this.sqlRecorder =  new SQLRecorder(MycatServer.getInstance().getConfig().getSystem().getSqlRecordCount());
+		this.sqlHighStat=new UserSqlHigh();
 	}
 
 	public String getUser() {
@@ -99,6 +101,11 @@ public class UserStat {
 		
 		//记录SQL
 		this.sqlStat.add(sql, executeTime, startTime, endTime );
+		
+		//记录高频SQL
+		this.sqlHighStat.addSql(sql, executeTime, startTime, endTime);
 	}
-
+	public  UserSqlHigh getSqlHigh(){
+		return this.sqlHighStat;
+	}
 }
